+++ src/components/modules/General/tmdb.js
@@ -1,13 +1,9 @@
 // TMDB stuff used
-//import store from '../../../store';
-//import { wtconfig } from './wtutils';
 import axios from 'axios';
-import { wtutils } from './wtutils';
+import {  wtconfig, wtutils } from './wtutils';
 
 const log = require('electron-log');
 const {JSONPath} = require('jsonpath-plus');
-
-
 
 const tmdb = new class TMDB {
     constructor() {
@@ -35,6 +31,8 @@ const tmdb = new class TMDB {
         const apiKey = wtutils.envVarLocal( 'Key_tmdb' );
         url = `${url}&api_key=${apiKey}`;
         const result = {};
+        let seasonCount = 0;
+        let episodeCount = 0;
         await axios({
             method: 'get',
             url: url,
@@ -43,13 +41,21 @@ const tmdb = new class TMDB {
             .then((response) => {
               log.debug('[tmdb.js] (getTMDBShowInfo) - Response from getTMDBShowInfo recieved');
               result['Status (Cloud)'] = JSONPath({ path: "$.status", json: response.data })[0];
-              result['Episode Count (Cloud)'] = JSONPath({ path: "$.number_of_episodes", json: response.data })[0];
-              result['Season Count (Cloud)'] = JSONPath({ path: "$.number_of_seasons", json: response.data })[0];
               // Now get season/episode
               const seasons = JSONPath({ path: "$..seasons[*]", json: response.data })
               let Seasons_Cloud = {};
               for ( var idx in seasons ){
-                Seasons_Cloud[JSONPath({ path: "$..season_number", json: seasons[idx]})] = JSONPath({ path: "$..episode_count", json: seasons[idx]})[0];
+                if ( JSONPath({ path: "$..season_number", json: seasons[idx]}) == 0) {
+                    if ( !wtconfig.get('ET.noSpecials') ){
+                        Seasons_Cloud[JSONPath({ path: "$..season_number", json: seasons[idx]})] = JSONPath({ path: "$..episode_count", json: seasons[idx]})[0];
+                        seasonCount++;
+                        episodeCount = episodeCount + JSONPath({ path: "$..episode_count", json: seasons[idx]})[0];
+                    }
+                } else {
+                    Seasons_Cloud[JSONPath({ path: "$..season_number", json: seasons[idx]})] = JSONPath({ path: "$..episode_count", json: seasons[idx]})[0];
+                    seasonCount++;
+                    episodeCount = episodeCount + JSONPath({ path: "$..episode_count", json: seasons[idx]})[0];
+                }
               }
               result['Seasons (Cloud)'] = Seasons_Cloud;
             })
@@ -75,6 +81,8 @@ const tmdb = new class TMDB {
                   return result;
               }
             })
+        result['Episode Count (Cloud)'] = episodeCount;
+        result['Season Count (Cloud)'] = seasonCount;
         log.silly(`[tmdb.js] (getTMDBShowInfo) - Returning: ${JSON.stringify(result)}`);
         return result;
     }